input: touchscreen: support Allwinner SoCs' touchscreen
This adds support for Allwinner SoCs' (A10, A13 and A31) resistive
touchscreen. This driver is probed by the MFD sunxi-gpadc-mfd.

This driver uses ADC channels exposed through the IIO framework by
sunxi-gpadc-iio to get its data. When opening this input device, it will
start buffering in the ADC driver and enable a TP_UP_PENDING irq. The ADC
driver will fill in a buffer with all data and call the callback the input
device associated with this buffer. The input device will then read the
buffer two by two and send X and Y coordinates to the input framework based
on what it received from the ADC's buffer. When closing this input device,
the buffering is stopped.

Note that locations in the first received buffer after an TP_UP_PENDING irq
occurred are unreliable, thus dropped.

Signed-off-by: Quentin Schulz <quentin.schulz@free-electrons.com>