[WIP] net: wireless: add ESP8089 WiFi driver
[WIP] Add the WiFi driver for the ESP8089.

This requires three firmwares (eagle_fw_ate_config_v19.bin,
eagle_fw_first_init_v19.bin and eagle_fw_second_init_v19.bin) which are
not in linux-firmware at the moment and will not be in the Linux kernel.

At the moment, this driver MUST be built as a module, or it won't work.

Signed-off-by: Quentin Schulz <quentin.schulz@free-electrons.com>